moved model_definitions in a separate file
